+++ plugins/drivers/imap.php
@@ -162,7 +162,8 @@ if (isset($_GET["imap"])) {
 	function db_collation(string $db, array $collations) {
 	}
 
-	function information_schema(string $db) {
+	function information_schema(string $db): bool {
+		return false;
 	}
 
 	function indexes(string $table, ?Db $connection2 = null): array {