+++ plugins/drivers/redis.php
@@ -430,7 +430,8 @@ if (isset($_GET["redis"])) {
 	function db_collation(string $db, array $collations) {
 	}
 
-	function information_schema(string $db) {
+	function information_schema(string $db): bool {
+		return false;
 	}
 
 	function indexes(string $table, ?Db $connection2 = null): array {